fix: type import structure
