+++ src/Components/QuizDetails/SideMenu/index.tsx
@@ -1,21 +1,8 @@
 import { Grid } from "@mui/material";
-import { QuizType } from "../../../types";
 import { QuizInfoSection } from "./QuizInfoSection";
 import { SourcesList } from "./SourcesList";
 import { useParams } from "react-router-dom";
 import { useGetQuiz } from "../../../Api/quizzes";
-
-export interface ResponseSourceType {
-  created_at: string;
-  file_hash: string;
-  file_name: string;
-  source_id: number;
-}
-
-export interface QuizInfoType {
-  quiz: QuizType;
-  sources: ResponseSourceType[];
-}
 
 export const InfoSideMenu = () => {
   const { quizId } = useParams();
